llm-proto-gen/requirements.txt

langchain
camelot-py[cv]
opencv-python
pypdf
PyPDF2==2.12.1